+++ src/app/guides/theming/theming.component.ts
@@ -1,53 +1,36 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Theme, ThemeService } from 'ngx-slice-kit';
 import { ApiDefinition } from '../../shared/model';
-import { ThemeService } from '../../../../libs/ngx-slice-kit/src/lib/core/theme/theme.service';
 import { LayoutService } from '../../shared/services';
+import { angularJsonExampleRef } from '../get-started/get-started.component';
+import { timer } from 'rxjs';
+import { defaultColors } from '../../../../libs/ngx-slice-kit/src/lib/core/theme/theme.model';
 
 @Component({
     selector: 'app-theming',
     templateUrl: './theming.component.html',
     styleUrls: ['./theming.component.scss', '../guides.module.scss']
 })
-export class ThemingComponent implements OnInit {
+export class ThemingComponent implements OnInit, OnDestroy {
 
     coreStylesExample = `@import './libs/ngx-slice-kit/src/lib/core/styles/core';`;
-    fontStylesExample = `@import './libs/ngx-slice-kit/src/lib/core/styles/typography';`;
-    colorPaletteExample = `class ColorPalette {
-    value: string;
-    text_value?: string;
-    background: string;
-    name: string;
-    text?: string;
-    prop?: string;
-    alpha?: number;
-    rgb?: string;
-    hex?: string;
-}`;
+    angularJson = angularJsonExampleRef;
 
     themesDefinitions: ApiDefinition[];
+    themes: Theme[] = [];
 
     constructor(
-        private themeService: ThemeService,
+        public themeService: ThemeService,
         private layoutService: LayoutService,
     ) {
     }
 
-    getCurrentThemesDefs(): ApiDefinition[] {
-        const currentTheme = this.themeService.themes.find(t => t.name === this.layoutService.themeName);
-        const themesDefinitions = [];
-        Object.keys(currentTheme).forEach(k => {
-            if (k !== 'name') {
-                themesDefinitions.push(new ApiDefinition({
-                    label: k,
-                    type: 'rgb-raw value',
-                    value: `rgb(${currentTheme[k]})`,
-                }));
-            }
-        });
-        return themesDefinitions;
+    ngOnInit(): void {
+        this.themes = this.themeService.themes;
+        // this.themes = [...this.themeService.themes, new Theme({name: 'custom', ...defaultColors})];
     }
 
-    ngOnInit(): void {
+    ngOnDestroy(): void {
     }
 
 }